Move files, instead of copy. Remove FlattenHierarchy from workflow.
