+++ workflow/src/main/java/bio/terra/tanagra/workflow/PrecomputeCounts.java
@@ -45,7 +45,7 @@ import org.apache.beam.sdk.values.PCollection;
  * <p>occurrences (node,what_to_count) = list of condition_occurrence rows (condition,person)
  *
  * <p>ancestor descendant relationships (ancestor,descendant) = list of ancestor-descendant pairs in
- * the condition hierarchy, as generated by the ({@link FlattenHierarchy}) workflow
+ * the condition hierarchy, as generated by the (FlattenHierarchy) workflow
  *
  * <p>OUTPUT: counts (node,count) = (condition, # people with >=1 occurrence of the condition or any
  * of its children)
